makefile

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: lrocigno <lrocigno@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/03 14:44:45 by lrocigno          #+#    #+#             */
-/*   Updated: 2021/02/05 14:48:16 by lrocigno         ###   ########.fr       */
+/*   Updated: 2021/02/18 11:29:55 by lrocigno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	char	*aux;
-	char	cc;
-	int		i;
+	char		*aux;
+	char		cc;
+	size_t		i;
 
 	aux = (char*)s;
 	cc = (char)c;
